@@ -1,43 +1,48 @@
 {
-    "name": "terminal-support-emoji",
-    "version": "1.0.0",
-    "description": "Node.js - Detect if the terminal is powershell",
-    "main": "index.js",
-    "scripts": {
-        "prepublishOnly": "pkg-ok",
-        "doc": "jsdoc -c ./jsdoc.json -r -R ./README.md -P ./package.json --verbose"
-    },
-    "repository": {
-        "type": "git",
-        "url": "git+https://github.com/fraxken/is-powershell.git"
-    },
-    "keywords": [
-        "is",
-        "powershell"
-    ],
-    "author": "GENTILHOMME Thomas <gentilhomme.thomas@gmail.com>",
-    "license": "MIT",
-    "bugs": {
-        "url": "https://github.com/fraxken/is-powershell/issues"
-    },
-    "homepage": "https://github.com/fraxken/is-powershell#readme",
-    "dependencies": {},
-    "devDependencies": {
-        "@commitlint/cli": "^7.2.1",
-        "@commitlint/config-conventional": "^7.1.2",
-        "@escommunity/minami": "^1.0.0",
-        "@slimio/eslint-config": "^2.0.3",
-        "@types/node": "^10.12.18",
-        "cross-env": "^5.2.0",
-        "eslint": "^5.10.0",
-        "husky": "^1.3.0",
-        "jsdoc": "^3.5.5",
-        "pkg-ok": "^2.3.1"
-    },
-    "engines": {
-        "node": ">=10"
-    },
-    "optionalDependencies": {
-        "is-powershell": "^1.1.1"
-    }
+  "name": "terminal-support-emoji",
+  "type": "module",
+  "version": "1.0.0",
+  "description": "Node.js - Detects if the terminal is supporting emoji",
+  "main": "index.js",
+  "scripts": {
+    "prepublishOnly": "pkg-ok",
+    "doc": "jsdoc -c ./jsdoc.json -r -R ./README.md -P ./package.json --verbose"
+  },
+  "repository": {
+    "type": "git",
+    "url": "https://github.com/TopCli/terminal-support-emoji.git"
+  },
+  "files": [
+    "index.js",
+    "index.d.ts"
+  ],
+  "keywords": [
+    "terminal",
+    "supports",
+    "emoji"
+  ],
+  "author": "GENTILHOMME Thomas <gentilhomme.thomas@gmail.com>",
+  "license": "MIT",
+  "bugs": {
+    "url": "https://github.com/TopCli/terminal-support-emoji/issues"
+  },
+  "homepage": "https://github.com/TopCli/terminal-support-emoji",
+  "devDependencies": {
+    "@commitlint/cli": "^15.0.0",
+    "@commitlint/config-conventional": "^15.0.0",
+    "@escommunity/minami": "^1.0.0",
+    "@slimio/eslint-config": "^5.0.0",
+    "@types/node": "^17.0.4",
+    "cross-env": "^7.0.2",
+    "eslint": "^8.3.0",
+    "husky": "^7.0.4",
+    "jsdoc": "^3.5.5",
+    "pkg-ok": "^2.3.1"
+  },
+  "engines": {
+    "node": ">=14"
+  },
+  "optionalDependencies": {
+    "is-powershell": "^1.1.1"
+  }
 }
