@@ -1,9 +1,10 @@
 // Require Node.js Dependencies
-const { spawnSync } = require("child_process");
+import { spawnSync } from "child_process";
 
 // CONSTANTS
 const BASH_CMD = "echo -e '\xE0\xA5\xA5' | wc -m";
-const POWERSHELL_CMD = "echo '\xE0\xA5\xA5' | measure -Character | Select-Object -expand Characters";
+const POWERSHELL_CMD =
+  "echo '\xE0\xA5\xA5' | measure -Character | Select-Object -expand Characters";
 
 /**
  * @func runSysCommand
@@ -11,14 +12,14 @@ const POWERSHELL_CMD = "echo '\xE0\xA5\xA5' | measure -Character | Select-Object
  * @returns {Promise<Number>}
  */
 function runSysCommand(command) {
-    return new Promise((resolve, reject) => {
-        const { error, stdout } = spawnSync("cmd", ["/c", command]);
-        if (error) {
-            return reject(error);
-        }
+  return new Promise((resolve, reject) => {
+    const { error, stdout } = spawnSync("cmd", ["/c", command]);
+    if (error) {
+      return reject(error);
+    }
 
-        return resolve(Number(stdout.toString()));
-    });
+    return resolve(Number(stdout.toString()));
+  });
 }
 
 /**
@@ -27,16 +28,16 @@ function runSysCommand(command) {
  * @desc Detect if the parent pid (terminal) support emoji or not!
  * @returns {Promise<Boolean>}
  */
-async function supportEmoji() {
-    if (process.platform === "win32") {
-        // eslint-disable-next-line
-        const isPowershell = require("is-powershell");
-        const termIsPowershell = await isPowershell();
+export async function supportEmoji() {
+  if (process.platform === "win32") {
+    // eslint-disable-next-line
+    const isPowershell = require("is-powershell");
+    const termIsPowershell = await isPowershell();
 
-        return await runSysCommand(termIsPowershell ? POWERSHELL_CMD : BASH_CMD) === 2;
-    }
+    return (
+      (await runSysCommand(termIsPowershell ? POWERSHELL_CMD : BASH_CMD)) === 2
+    );
+  }
 
-    return await runSysCommand(BASH_CMD) === 2;
+  return (await runSysCommand(BASH_CMD)) === 2;
 }
-
-module.exports = supportEmoji;
