@@ -1,3 +1,3 @@
-const supportEmoji = require("./index");
+import { supportEmoji } from "./index.js";
 
 supportEmoji().then(console.log).catch(console.error);
